:art: move register to a single module
